requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-cleanup==6.0.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-import-export==3.0.1
django-multiselectfield==0.1.12
django-shortcuts==1.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.1.1
gunicorn==20.1.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python-dotenv==0.21.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
tablib==3.2.1
tomli==2.0.1
traitlets==5.4.0
typing_extensions==4.4.0
unicodecsv==0.14.1
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
